RAD-209 Extract hl7snd into own class
* extract sending of HL7 into own class HL7Sender
so DicomUtils concern is Dicom stuff
